requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
setuptools
pymongo[srv]==4.0
scikit-learn
# -e .

